requirements.txt
selenium==4.1.0
pyyaml==6.0
python-dotenv==0.19.2